+++ src/api/dto/CalloutResponseDto.ts
@@ -31,7 +31,9 @@ import { GetCalloutDto } from "@api/dto/CalloutDto";
 import { GetCalloutResponseCommentDto } from "@api/dto/CalloutResponseCommentDto";
 import { GetCalloutTagDto } from "@api/dto/CalloutTagDto";
 
-import Callout, { CalloutResponseViewSchema } from "@models/Callout";
+import Callout from "@models/Callout";
+
+import { CalloutResponseViewSchema } from "@type/callout-response-view-schema";
 
 export interface BaseGetCalloutResponseOptsDto {
   callout?: Callout;
