+++ src/type/callout-map-schema.ts
@@ -1,0 +1,12 @@
+export interface CalloutMapSchema {
+  style: string;
+  center: [number, number];
+  bounds: [[number, number], [number, number]];
+  minZoom: number;
+  maxZoom: number;
+  initialZoom: number;
+  addressProp: string;
+  addressPattern: string;
+  addressPatternProp: string;
+  geocodeCountries?: string;
+}
